asyncio worker framework for cascading geocoder
This is the outline of how you might implement an asyncio-worker based
cascading geocoding pipeline.

The tricky bit is how to do the batching in a reasonable way. Here,
there's a generic "abatch" worker that batches and then puts on the next
queue, but it could be simplifed by rolling that into the geocoder code
itself. Probably makes sense because worker communication is always
weird.